Added Json users file, custom validator checking on login if id and password match user's, CGU checkbox must be checked for sign up
