+++ src/app/game-list/game-list.component.css
@@ -15,6 +15,11 @@ h1 {
   margin-bottom: 2rem;
 }
 
+button {
+  display: block;
+  margin: auto;
+}
+
 @media screen and (min-width: 768px) {
   app-game-card:hover {
     transform: scale(1.1);
