+++ src/app/login/login.component.html
@@ -1,5 +1,5 @@
 <div id="login-page">
-  <h1>Se connecter</h1>
+  <h1>{{this.title}}</h1>
   <form
     (ngSubmit)="login_form.valid && onSubmit()"
     [formGroup]="loginForm"
@@ -8,7 +8,7 @@
     <div class="input-field">
       <label for="id">Identifiant: </label><br />
       <input type="text" name="id" formControlName="id" /><br />
-      <div *ngIf="loginForm.controls.id.errors && login_form.submitted">
+      <div *ngIf="loginForm.controls.id.errors && login_form.submitted && !userLoggedIn">
         <span *ngIf="loginForm.controls.id.errors?.['required']"
           >Le champ Identifiant doit être rempli</span
         >
@@ -18,11 +18,15 @@
     <div class="input-field">
       <label for="password">Mot de passe: </label><br />
       <input type="password" name="password" formControlName="password" /><br />
-      <div *ngIf="loginForm.controls.password.errors && login_form.submitted">
+      <div *ngIf="loginForm.controls.password.errors && login_form.submitted && !userLoggedIn">
         <span *ngIf="loginForm.controls.password.errors?.['required']"
-          >Le champ Identifiant doit être rempli</span
+          >Le champ mot de passe doit être rempli</span
         >
       </div>
+    </div>
+
+    <div *ngIf="loginForm.errors && login_form.submitted && !userLoggedIn">
+      <span *ngIf="loginForm.errors?.['wrongLogin']">Identifiant ou mot de passe incorrect</span>
     </div>
 
     <button type="submit">Connexion</button>
