+++ src/app/usersJson.json
@@ -1,0 +1,4 @@
+{
+    "id": "Lucas",
+    "password": "CoucouLucas"
+}
